# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para um sistema de microsserviços que gerencia filmes.",
        "title": "API de Gerenciamento de Filmes",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Henrique Alencar",
            "url": "https://github.com/alenrique",
            "email": "henriquealencardev@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/movies": {
            "get": {
                "description": "Retorna uma lista com todos os filmes cadastrados no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Filmes"
                ],
                "summary": "Lista todos os filmes",
                "responses": {
                    "200": {
                        "description": "Lista de filmes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.MovieSwagger"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Adiciona um novo filme à coleção a partir dos dados enviados no corpo da requisição.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Filmes"
                ],
                "summary": "Cria um novo filme",
                "parameters": [
                    {
                        "description": "Dados do Filme para Criar",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateMovieRequestSwagger"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Filme criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/main.MovieSwagger"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "Retorna os detalhes de um filme específico com base no seu ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Filmes"
                ],
                "summary": "Busca um filme por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Filme",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filme encontrado",
                        "schema": {
                            "$ref": "#/definitions/main.MovieSwagger"
                        }
                    },
                    "404": {
                        "description": "Filme não encontrado",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um filme da coleção com base no seu ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Filmes"
                ],
                "summary": "Deleta um filme por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Filme a ser deletado",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Filme deletado com sucesso (sem conteúdo de resposta)"
                    },
                    "404": {
                        "description": "Filme não encontrado",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CreateMovieRequestSwagger": {
            "type": "object",
            "properties": {
                "director": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "main.MovieSwagger": {
            "type": "object",
            "properties": {
                "director": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    }
}